refactor: piece 패키지 정리
