+++ src/main/java/domain/piece/PieceColor.java
@@ -1,5 +1,13 @@
 package domain.piece;
 
 public enum PieceColor {
-    BLACK, WHITE
+    BLACK, WHITE;
+
+    public PieceColor toggle() {
+        if (this == BLACK) {
+            return WHITE;
+        }
+
+        return BLACK;
+    }
 }
