+++ src/main/java/domain/piece/PieceType.java
@@ -1,0 +1,10 @@
+package domain.piece;
+
+public enum PieceType {
+    PAWN,
+    ROOK,
+    KNIGHT,
+    BISHOP,
+    QUEEN,
+    KING
+}
